+++ src/elements/actions/DeleteProduct.php
@@ -1,0 +1,39 @@
+<?php
+/**
+ * @link https://craftcms.com/
+ * @copyright Copyright (c) Pixel & Tonic, Inc.
+ * @license https://craftcms.github.io/license/
+ */
+
+namespace craft\digitalproducts\elements\actions;
+
+use Craft;
+use craft\elements\actions\Delete;
+use craft\elements\db\ElementQueryInterface;
+
+/**
+ * Delete Product Action
+ *
+ * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
+ * @since 2.4
+ */
+class DeleteProduct extends Delete
+{
+    /**
+     * @inheritdoc
+     */
+    public function performAction(ElementQueryInterface $query = null): bool
+    {
+        if (!$query) {
+            return false;
+        }
+
+        foreach ($query->all() as $product) {
+            Craft::$app->getElements()->deleteElement($product);
+        }
+
+        $this->setMessage(Craft::t('digital-products', 'Products deleted.'));
+
+        return true;
+    }
+}